Create: 포스트 발행 페이지 레이아웃 수정 중
이미지 선택 컴포넌트 분리해서 별도 페이지 구성
